Create README.md
Despite McDonald's health revolution, the chain still cannot say that the dishes are healthy, but it is effective for information. So we decided to add a column to the stat that would determine for each dish whether it is harmful or not, we conducted research for each and every dish from the menu in the status and with the help of the features given in it and various opinions of professionals on the Internet we determined for each dish whether it is harmful to health or not.
This column will form the ranking with the help of the built model we will know how to classify each new dose we add to our data whether it is harmful or not.